Update component library
