+++ Gemfile
@@ -31,12 +31,12 @@ group :development do
   # Display performance information such as SQL time and flame graphs for each request in your browser.
   # Can be configured to work on production as well see: https://github.com/MiniProfiler/rack-mini-profiler/blob/master/README.md
   gem 'rack-mini-profiler', '~> 2.0'
-  gem 'debug'
 end
 
 group :development, :test do
   gem 'rspec-rails'
   gem 'capybara'
+  gem 'debug'
   gem 'selenium-webdriver', '!= 3.13.0'
   gem 'factory_bot_rails'
   gem 'database_cleaner'
