--- app/components/blacklight/top_navbar_component.html.erb
@@ -1,189 +1,0 @@
-<div class="container">
-  <a class="su-brand-bar__logo" href="https://www.stanford.edu"
-    >Stanford University</a
-  >
-</div>
-<nav
-  class="navbar navbar-expand-md bg-dark palo-alto-base"
-  role="navigation"
->
-  <div class="container">
-    <a
-      class="mb-0 navbar-brand navbar-logo"
-      href="https://library.stanford.edu"
-      >Stanford University Libraries</a
-    >
-    <button
-      class="navbar-toggler navbar-toggler-right collapsed"
-      type="button"
-      data-toggle="collapse"
-      data-bs-toggle="collapse"
-      data-target="#user-util-collapse"
-      data-bs-target="#user-util-collapse"
-      aria-controls="user-util-collapse"
-      aria-expanded="false"
-      aria-label="Toggle navigation"
-    >
-      <span class="navbar-toggler-icon"></span>
-    </button>
-
-    <div class="d-md-block d-none">
-      <ul class="navbar-nav">
-        <li class="nav-item">
-          <%= link_to bookmarks_path, id:'bookmarks_nav', class: 'nav-link' do %>
-            <%= t('blacklight.header_links.bookmarks') %>
-            (<span data-role='bookmark-counter'><%= helpers.current_or_guest_user.bookmarks.count %></span>)
-          <% end %>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" href="/search_history">Search history</a>
-        </li>
-        <li class="nav-item">
-          <a
-            href="#feedback"
-            class="nav-link"
-            data-bs-toggle="collapse"
-            data-bs-target="#feedback"
-            >Feedback</a
-          >
-        </li>
-        <li class="nav-item dropdown">
-          <% if helpers.current_user && !helpers.current_user.sunet.blank? %>
-            <button class="nav-link btn-link logout-dropdown dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">
-              Logged in: <%= helpers.current_user.display_name ||  helpers.current_user.sunet %>
-            </button>
-            <ul class="dropdown-menu">
-              <li><%= link_to "Logout", destroy_user_session_path, class: "dropdown-item" %></li>
-            </ul>
-          <% else %>
-            <%= link_to 'Login', new_user_session_path(referrer: request.original_url), class: "nav-link" %>
-          <% end %>
-        </li>
-      </ul>
-    </div>
-  </div>
-</nav>
-
-<div id="feedback" class="collapse">
-  <div class="container pt-3">
-    <div class="row justify-content-center">
-      <form
-        class="col-md-8 feedback-form"
-        action="/feedback"
-        accept-charset="UTF-8"
-        method="post"
-      >
-        <input type="hidden" id="url" name="url" value="<%= request.original_url %>" />
-        <div class="alert alert-info" role="alert">
-          Reporting from:
-          <%= root_url %>
-        </div>
-        <div class="mb-3 row">
-          <label class="col-md-3 col-form-label" for="message"
-            >Message</label
-          >
-          <div class="col-md-9">
-            <textarea
-              class="form-control"
-              rows="5"
-              required="required"
-              data-form-type="other"
-              name="message"
-              id="message"
-            ></textarea>
-          </div>
-        </div>
-        <div class="mb-3 row">
-          <label class="col-md-3 col-form-label" for="name"
-            >Your name</label
-          >
-          <div class="col-md-9">
-            <%= text_field_tag :name, helpers.current_user&.display_name, autocomplete: 'name', class:"form-control", required: true %>
-          </div>
-        </div>
-        <div class="mb-3 row">
-          <label class="col-md-3 col-form-label" for="email"
-            >Your email</label
-          >
-          <div class="col-md-9">
-            <%= email_field_tag :email, helpers.current_user&.email, autocomplete: 'email', class:"form-control", required: true %>
-          </div>
-        </div>
-        <div class="mb-3 row">
-          <div class="col-md-9 offset-md-3">
-            <% if helpers.current_user.blank? %>
-              <%= helpers.recaptcha_tags %>
-
-              <p>(Stanford users can avoid this Captcha by logging in.)</p>
-            <% end %>
-          </div>
-        </div>
-        <div class="mb-3 row">
-          <div class="col-md-9 offset-md-3">
-            <input
-              type="submit"
-              name="commit"
-              value="Send"
-              class="btn btn-secondary"
-              data-disable-with="Send"
-            />
-            <button
-              name="button"
-              type="button"
-              data-bs-toggle="collapse"
-              data-bs-target="#feedback"
-              class="btn btn-outline-secondary"
-            >
-              Cancel
-            </button>
-          </div>
-        </div>
-      </form>
-    </div>
-  </div>
-</div>
-<div class="masthead bg-dark palo-alto-base">
-  <div class="container">
-    <div class="row align-items-center">
-      <div class="col-md-8 d-flex">
-        <div class="h1 my-3"><a href="/">EarthWorks</a></div>
-      </div>
-      <div
-        class="navbar collapse navbar-collapse ms-3"
-        aria-label="browse" id="user-util-collapse">
-        <ul class="navbar-nav">
-          <li class="nav-item d-md-none">
-            <%= link_to bookmarks_path, id:'bookmarks_nav', class: 'nav-link' do %>
-              <%= t('blacklight.header_links.bookmarks') %>
-              (<span data-role='bookmark-counter'><%= helpers.current_or_guest_user.bookmarks.count %></span>)
-            <% end %>
-          </li>
-          <li class="nav-item d-md-none">
-            <a class="nav-link" href="/search_history">Search history</a>
-          </li>
-          <li class="nav-item d-md-none">
-            <a
-              href="#feedback"
-              class="nav-link"
-              data-bs-toggle="collapse"
-              data-bs-target="#feedback"
-              >Feedback</a
-            >
-          </li>
-          <li class="nav-item dropdown d-md-none">
-            <% if helpers.current_user && !helpers.current_user.sunet.blank? %>
-              <button class="nav-link btn-link logout-dropdown dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">
-                Logged in: <%= helpers.current_user.display_name ||  helpers.current_user.sunet %>
-              </button>
-              <ul class="dropdown-menu">
-                <li><%= link_to "Logout", destroy_user_session_path, class: "dropdown-item" %></li>
-              </ul>
-            <% else %>
-              <%= link_to 'Login', new_user_session_path(referrer: request.original_url), class: "nav-link" %>
-            <% end %>
-          </li>
-        </ul>
-      </div>
-    </div>
-  </div>
-</div>
