--- app/components/geoblacklight/header_component.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Geoblacklight
-  class HeaderComponent < Blacklight::HeaderComponent
-    def landing_page?
-      current_page?(root_path) && !helpers.has_search_parameters?
-    end
-  end
-end
