+++ app/components/header_component.html.erb
@@ -1,5 +1,6 @@
 <header>
   <%= top_bar %>
+
   <% unless landing_page? %>
     <%= search_bar %>
   <% end %>
