+++ app/components/header_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class HeaderComponent < Blacklight::HeaderComponent
+  def landing_page?
+    current_page?(root_path) && !helpers.has_search_parameters?
+  end
+end
