+++ app/components/top_navbar_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class TopNavbarComponent < Blacklight::TopNavbarComponent
+end
