+++ app/views/layouts/blacklight/base.html.erb
@@ -1,13 +1,31 @@
 <!DOCTYPE html>
+
 <%= content_tag :html, class: 'no-js', **html_tag_attributes do %>
   <head>
     <meta charset="utf-8">
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="google-site-verification" content="<%= Settings.GOOGLE_SITE_VERIFICATION %>" />
-    <meta name="msvalidate.01" content="<%= Settings.BING_SITE_VERIFICATION %>" />
+
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1, shrink-to-fit=no"
+    >
+
+    <meta
+      name="google-site-verification"
+      content="<%= Settings.GOOGLE_SITE_VERIFICATION %>"
+    />
+
+    <meta
+      name="msvalidate.01"
+      content="<%= Settings.BING_SITE_VERIFICATION %>"
+    />
+
     <!-- Google tag (gtag.js) -->
-    <script async src="https://www.googletagmanager.com/gtag/js?id=G-ZL7MQNBFW9"></script>
+    <script
+      async
+      src="https://www.googletagmanager.com/gtag/js?id=G-ZL7MQNBFW9"
+    ></script>
+
     <script>
       window.dataLayer = window.dataLayer || [];
       function gtag(){ dataLayer.push(arguments); }
@@ -22,13 +40,29 @@
     </script>
 
     <title><%= render_page_title %></title>
+
     <script>
       document.querySelector('html').classList.remove('no-js');
     </script>
+
     <%= opensearch_description_tag application_name, opensearch_catalog_url(format: 'xml') %>
-    <link rel="icon" href="https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-01-24/styles/icon.png" type="image/png">
-    <link rel="icon" href="https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-01-24/styles/icon.svg" type="image/svg+xml">
-    <link rel="apple-touch-icon" href="https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-01-24/styles/icon.png">
+
+    <link
+      rel="icon"
+      href="https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-09-11/styles/icon.png"
+      type="image/png"
+    >
+
+    <link
+      rel="icon"
+      href="https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-09-11/styles/icon.svg"
+      type="image/svg+xml"
+    >
+
+    <link
+      rel="apple-touch-icon"
+      href="https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-09-11/styles/icon.png"
+    >
 
     <!-- viewer styles (non-render-blocking; deferred) -->
     <%= preload_link_tag "https://cdn.skypack.dev/leaflet@1.9.4/dist/leaflet.css", as: "style", onload: "this.onload=null;this.rel='stylesheet'" %>
@@ -36,9 +70,9 @@
     <%= preload_link_tag "https://cdn.skypack.dev/ol@8.1.0/ol.css", as: "style", onload: "this.onload=null;this.rel='stylesheet'" %>
 
     <!-- main app styles -->
-    <%= stylesheet_link_tag "geoblacklight", media: "all", "data-turbo-track": "reload"  %>
-    <%= stylesheet_link_tag "earthworks", media: "all", "data-turbo-track": "reload"  %>
-    <%= stylesheet_link_tag "https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-01-10/styles/sul.css" %>
+    <%= stylesheet_link_tag "geoblacklight", media: "all", "data-turbo-track": "reload" %>
+    <%= stylesheet_link_tag "earthworks", media: "all", "data-turbo-track": "reload" %>
+    <%= stylesheet_link_tag "https://cdn.jsdelivr.net/gh/sul-dlss/component-library@v2025-09-11/styles/sul.css" %>
     <%= stylesheet_link_tag "https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css" %>
 
     <!-- scripts -->
@@ -57,14 +91,22 @@
     <%= csrf_meta_tags %>
     <%= content_for(:head) %>
   </head>
+
   <body class="<%= render_body_class %>">
     <%= render blacklight_config.skip_link_component.new do %>
       <%= content_for(:skip_links) %>
     <% end %>
 
-    <%= render blacklight_config.header_component.new(blacklight_config: blacklight_config) %>
+    <%= render blacklight_config.header_component.new(blacklight_config: blacklight_config) do |c| %>
+      <%= c.with_top_bar(component: TopNavbarComponent) %>
+    <% end %>
 
-    <main id="main-container" class="<%= container_classes %>" role="main" aria-label="<%= t('blacklight.main.aria.main_container') %>">
+    <main
+      id="main-container"
+      class="<%= container_classes %>"
+      role="main"
+      aria-label="<%= t('blacklight.main.aria.main_container') %>"
+    >
       <%= content_for(:container_header) %>
 
       <%= render partial: 'shared/flash_msg', layout: 'shared/flash_messages' %>
@@ -74,7 +116,7 @@
       </div>
     </main>
 
-    <%= render partial: 'shared/footer' %>
-    <%= render partial: 'shared/modal' %>
+    <%= render 'shared/footer' %>
+    <%= render 'shared/modal' %>
   </body>
 <% end %>
